# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cabales API
  version: 1.2.0
  license: { name: ISC, identifier: ISC }
  description: >-
    API de autenticación, grupos, eventos, gastos exactos, liquidaciones, fondos, presupuestos, documentos, OCR, Cabudas, estadísticas, avisos, logros y privacidad. Importes en centavos.
    Todas las respuestas /api/v1 usan Cache-Control private, no-store; health y ready usan no-store.
servers:
  - url: /
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Registra un usuario y emite sesión y CSRF
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201': { $ref: '#/components/responses/AuthSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Inicia una sesión opaca revocable
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200': { $ref: '#/components/responses/AuthSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoca la sesión actual
      operationId: logout
      security: [{ sessionCookie: [], csrfCookie: [], csrfHeader: [] }]
      responses:
        '200': { $ref: '#/components/responses/MutationSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Recupera identidad y CSRF validado desde su cookie
      operationId: getMe
      security: [{ sessionCookie: [], csrfCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/AuthSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    patch:
      tags: [Auth]
      summary: 'Rectifica nombre visible o idioma del titular'
      operationId: updateMe
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfilePatch' }
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /api/v1/auth/email-verification/request:
    post:
      tags: [Auth]
      summary: 'Solicita verificación de correo sin revelar si la cuenta existe'
      operationId: requestEmailVerification
      description: 'Límite por IP y por hash de correo; enfriamiento de 60 s y el token anterior queda invalidado. El enlace lleva el token en el fragmento (#token=) para que no llegue a logs del servidor.'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailInput' }
      responses:
        '202':
          description: 'Aceptada; respuesta idéntica exista o no la cuenta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/auth/email-verification/resend:
    post:
      tags: [Auth]
      summary: 'Reenvía verificación de correo sin revelar si la cuenta existe'
      operationId: resendEmailVerification
      description: 'Límite por IP y por hash de correo; enfriamiento de 60 s y el token anterior queda invalidado. El enlace lleva el token en el fragmento (#token=) para que no llegue a logs del servidor.'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailInput' }
      responses:
        '202':
          description: 'Aceptada; respuesta idéntica exista o no la cuenta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/auth/password-recovery/request:
    post:
      tags: [Auth]
      summary: 'Solicita recuperación de contraseña sin revelar si la cuenta existe'
      operationId: requestPasswordRecovery
      description: 'Límite por IP y por hash de correo; enfriamiento de 60 s y el token anterior queda invalidado. El enlace lleva el token en el fragmento (#token=) para que no llegue a logs del servidor.'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailInput' }
      responses:
        '202':
          description: 'Aceptada; respuesta idéntica exista o no la cuenta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/auth/password-recovery/resend:
    post:
      tags: [Auth]
      summary: 'Reenvía recuperación de contraseña sin revelar si la cuenta existe'
      operationId: resendPasswordRecovery
      description: 'Límite por IP y por hash de correo; enfriamiento de 60 s y el token anterior queda invalidado. El enlace lleva el token en el fragmento (#token=) para que no llegue a logs del servidor.'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EmailInput' }
      responses:
        '202':
          description: 'Aceptada; respuesta idéntica exista o no la cuenta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AcceptedEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/v1/auth/email-verification/confirm:
    post:
      tags: [Auth]
      summary: 'Confirma el correo con un token de un solo uso'
      operationId: confirmEmailVerification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TokenInput' }
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/auth/password-recovery/confirm:
    post:
      tags: [Auth]
      summary: 'Restablece la contraseña, revoca todas las sesiones y limpia cookies'
      operationId: confirmPasswordRecovery
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordResetInput' }
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: object
                    required: [reset]
                    properties:
                      reset:
                        const: true
        '400': { $ref: '#/components/responses/BadRequest' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
  /api/v1/groups/{groupId}/budgets:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Budgets]
      summary: 'Presupuestos con progreso del periodo actual'
      operationId: listBudgets
      parameters:
        - name: at
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: 'Fecha de referencia para el periodo actual; por defecto ahora.'
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Budget' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    post:
      tags: [Budgets]
      summary: 'Crea un presupuesto (propietario o administrador)'
      operationId: createBudget
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BudgetInput' }
      responses:
        '201':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/Budget' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/groups/{groupId}/budgets/{budgetId}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/BudgetId'
    get:
      tags: [Budgets]
      summary: 'Detalle con historial de los últimos 6 periodos'
      operationId: getBudget
      parameters:
        - name: at
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: 'Fecha de referencia para el periodo actual; por defecto ahora.'
      security:
        - sessionCookie: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/BudgetDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    patch:
      tags: [Budgets]
      summary: 'Edita un presupuesto'
      operationId: updateBudget
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BudgetPatch' }
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    const: true
                  data: { $ref: '#/components/schemas/Budget' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      tags: [Budgets]
      summary: 'Elimina un presupuesto'
      operationId: deleteBudget
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      responses:
        '200':
          description: 'Operación correcta'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MutationEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cabales_session
      description: Nombre predeterminado configurable mediante COOKIE_NAME.
    csrfCookie:
      type: apiKey
      in: cookie
      name: cabales_session_csrf
      description: Cookie de doble envío cuyo prefijo sigue COOKIE_NAME; /auth/me valida su hash antes de devolverla.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Debe coincidir con csrfCookie y con el hash ligado a la sesión.
  parameters:
    GroupId: { name: groupId, in: path, required: true, schema: { type: string, format: uuid } }
...
    BudgetId:
      name: budgetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    AuthSuccess:
      description: Identidad y CSRF
      headers:
        Cache-Control: { schema: { type: string, const: 'private, no-store' } }
        Set-Cookie:
          description: Register/login emiten sesión HttpOnly y CSRF legible; /me no rota cookies.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AuthEnvelope' }
    MutationSuccess:
      {
        description: Mutación completada,
        content:
          { application/json: { schema: { $ref: '#/components/schemas/MutationEnvelope' } } },
      }
...
    BadRequest:
      {
        description: Entrada inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Unauthorized:
      {
        description: Sesión ausente o inválida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Forbidden:
      {
        description: CSRF o permiso inválido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    NotFound:
      {
        description: Recurso no visible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    Conflict:
      {
        description: Estado concurrente o idempotencia incompatible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
    UnprocessableEntity:
      {
        description: Invariante de negocio incumplida,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
    TooManyRequests:
      {
        description: Límite de solicitudes excedido,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
...
    ServiceUnavailable:
      {
        description: PostgreSQL no disponible,
        content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } },
      }
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
...
    MutationEnvelope:
      type: object
      required: [success, data]
      properties:
        {
          success: { const: true },
          data: { type: object, additionalProperties: { type: boolean } },
        }
    User:
      type: object
      additionalProperties: false
      required: [id, email, displayName, avatarUrl, locale, emailVerified]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        displayName: { type: string }
        avatarUrl: { type: [string, 'null'], format: uri }
        locale: { type: string }
        emailVerified: { type: boolean }
    AuthEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { const: true }
        data:
          type: object
          additionalProperties: false
          required: [user, csrfToken]
          properties: { user: { $ref: '#/components/schemas/User' }, csrfToken: { type: string } }
    RegisterInput:
      type: object
      additionalProperties: false
      required: [email, password, displayName]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128 }
        displayName: { type: string, minLength: 2, maxLength: 120 }
    LoginInput:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        password: { type: string, minLength: 12, maxLength: 128 }
...
    AcceptedEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          const: true
        data:
          type: object
          required: [accepted]
          properties:
            accepted:
              const: true
    EmailInput:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
    TokenInput:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 256
          pattern: '^[A-Za-z0-9_-]+$'
    PasswordResetInput:
      type: object
      additionalProperties: false
      required: [token, password]
      properties:
        token:
          type: string
          minLength: 32
          maxLength: 256
          pattern: '^[A-Za-z0-9_-]+$'
        password:
          type: string
          minLength: 12
          maxLength: 128
    ProfilePatch:
      type: object
      additionalProperties: false
      required: []
      properties:
        displayName:
          type: string
          minLength: 2
          maxLength: 120
        locale:
          enum: [es, en]
      minProperties: 1
    UserEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          const: true
        data:
          type: object
          additionalProperties: false
          required: [user]
          properties:
            user: { $ref: '#/components/schemas/User' }
...
    Budget:
      type: object
      required:
        [
          id,
          groupId,
          name,
          amountCents,
          currency,
          period,
          startsAt,
          endsAt,
          alertThresholdPercent,
          createdAt,
          updatedAt,
          category,
          current,
        ]
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        name:
          type: string
        amountCents:
          type: integer
        currency:
          type: string
        period:
          enum: [WEEKLY, MONTHLY, YEARLY, CUSTOM]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: [string, 'null']
          format: date-time
        alertThresholdPercent:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          type: [object, 'null']
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            color:
              type: [string, 'null']
        current:
          type: object
          required:
            [
              periodStart,
              periodEnd,
              periodState,
              spentCents,
              expenseCount,
              remainingCents,
              overspentCents,
              progressPercent,
              alert,
            ]
          properties:
            periodStart:
              type: string
              format: date-time
            periodEnd:
              type: string
              format: date-time
            periodState:
              enum: [UPCOMING, ACTIVE, ENDED]
            spentCents:
              type: integer
            expenseCount:
              type: integer
            remainingCents:
              type: integer
            overspentCents:
              type: integer
            progressPercent:
              type: number
            alert:
              enum: [OK, WARNING, EXCEEDED]
    BudgetDetail:
      type: object
      required:
        [
          id,
          groupId,
          name,
          amountCents,
          currency,
          period,
          startsAt,
          endsAt,
          alertThresholdPercent,
          createdAt,
          updatedAt,
          category,
          current,
          history,
        ]
      properties:
        id:
          type: string
          format: uuid
        groupId:
          type: string
          format: uuid
        name:
          type: string
        amountCents:
          type: integer
        currency:
          type: string
        period:
          enum: [WEEKLY, MONTHLY, YEARLY, CUSTOM]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: [string, 'null']
          format: date-time
        alertThresholdPercent:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          type: [object, 'null']
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            color:
              type: [string, 'null']
        current:
          type: object
          required:
            [
              periodStart,
              periodEnd,
              periodState,
              spentCents,
              expenseCount,
              remainingCents,
              overspentCents,
              progressPercent,
              alert,
            ]
          properties:
            periodStart:
              type: string
              format: date-time
            periodEnd:
              type: string
              format: date-time
            periodState:
              enum: [UPCOMING, ACTIVE, ENDED]
            spentCents:
              type: integer
            expenseCount:
              type: integer
            remainingCents:
              type: integer
            overspentCents:
              type: integer
            progressPercent:
              type: number
            alert:
              enum: [OK, WARNING, EXCEEDED]
        history:
          type: array
          items:
            type: object
            required: [periodStart, periodEnd, spentCents, expenseCount]
            properties:
              periodStart:
                type: string
                format: date-time
              periodEnd:
                type: string
                format: date-time
              spentCents:
                type: integer
              expenseCount:
                type: integer
    BudgetInput:
      type: object
      additionalProperties: false
      required: [name, amountCents, period, startsAt]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        amountCents:
          type: integer
          minimum: 1
          maximum: 2147483647
        period:
          enum: [WEEKLY, MONTHLY, YEARLY, CUSTOM]
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: [string, 'null']
          format: date-time
          description: 'Obligatoria para CUSTOM y posterior a startsAt.'
        categoryId:
          type: [string, 'null']
          format: uuid
        alertThresholdPercent:
          type: integer
          minimum: 1
          maximum: 100
          default: 80
    BudgetPatch:
      type: object
      additionalProperties: false
      required: []
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        amountCents:
          type: integer
          minimum: 1
          maximum: 2147483647
        endsAt:
          type: [string, 'null']
          format: date-time
        categoryId:
          type: [string, 'null']
          format: uuid
        alertThresholdPercent:
          type: integer
          minimum: 1
          maximum: 100
      minProperties: 1
...